Load data from JSON and directly store devolution
Instead of using resources, it now loads a JSON file for the pokemon
data. This makes it more structured, as the data is not distributed
across resources. It also adds two new fields, the actual number and the
devolution forms, which means that the list isn't continous anymore and
that it is possible to determine which form evolves into which. This is
necessary as each Burmy form has two evolution paths. Previously it
assumed that forms only evolved into the same form, which is untrue.

Now `PokemonBase` also directly stores the devolution object, removing
the need to use `PokeInfoCalculator` to get the devolution base. The
same was done in `Pokemon` itself and the evolutions are also determined
once.

In the future "helper" methods in `PokeInfoCalculator` could be moved
into `Pokemon` or `PokemonBase` like it already does for
`Pokemon.getEvolutionLine`.